+++ changelog.md
@@ -1,17 +1,24 @@
 # Changelog
 
-## [1.12.18] - 2026-7-17 Rootelo Links
+## [1.12.18] - 2026-7-20 Rootelo Links
 
 ### New Features
 - dwd/profile/mrmirz-4412 redirects to profile/mrmirz
+- rejected posts are no longer deleted and display rejection notice for resubmission guidance
+- Option to be notified for new posts and when a post is marked as stable
 
 ### Improvements
 - profile and post name on detail pages now wraps and dynamically resizes
+- removed logged in check when viewing profiles
+- added cannonical dwd username field from the RDL API
+- added cached winrates by platform
+- /stats command shows threshold and defaults to official factions only
 
 ### Bug Fixes
 - profile detail page was crashing blank box score chart due to rename
 - message for submitted posts should now always link to the pending submittal page
 - discarded captain no longer cleared when editing a game
+- Leaderboard now reliably uses cached values when blank queries are given
 
 ## [1.12.17] - 2026-7-16 More RTM Feedback